Ld file parsing is completed. Next is to compse mesh of brick.
