+++ Assets/Scenes/MeshExporter/Scripts/LdFileParser.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 52ce28105556f8247b9c9d31fc456ce6
-timeCreated: 1509544133
+guid: a24b22610ed4dd549b0b83eb525298d9
+timeCreated: 1511679302
 licenseType: Free
 MonoImporter:
   externalObjects: {}
